requirements.txt
absl-py==0.2.2
astor==0.7.1
audioread==2.1.6
bleach==1.5.0
certifi==2016.2.28
decorator==4.3.0
gast==0.2.0
grpcio==1.13.0
html5lib==0.9999999
joblib==0.12.0
librosa==0.6.1
llvmlite==0.24.0
Markdown==2.6.11
numba==0.39.0
numpy==1.14.5
packaging==17.1
protobuf==3.6.0
pyparsing==2.2.0
python-speech-features==0.6
resampy==0.2.1
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
tensorboard==1.6.0
tensorflow-gpu==1.6.0
termcolor==1.1.0
webrtcvad==2.0.10
Werkzeug==0.14.1
